fw/music: fix progress bar overflow on long tracks
The rectangular now-playing layout computed the progress bar percent
as track_pos * 100 in 32-bit arithmetic. Once the playback position
exceeds ~11.9 h (UINT32_MAX / 100 ms), the product wraps and the bar
fill collapses (e.g. 9% shown at 13:20:55 of a 14:50:53 audiobook).

Promote the multiplication to 64-bit, as the round layout already does.

Fixes FIRM-4373

+++ src/fw/apps/system/music.c
@@ -936,7 +936,7 @@ static void prv_update_track_progress(MusicAppData *data) {
   } else {
     unsigned int percent = 0;
     if (data->track_length > 0) {
-      percent = MIN((data->track_pos * 100) / data->track_length, 100);
+      percent = MIN((unsigned int)(((uint64_t)data->track_pos * 100) / data->track_length), 100);
     }
     progress_layer_set_progress(&data->track_pos_bar, percent);
     prv_copy_time_period(data->position_buffer, sizeof(data->position_buffer),
